realpath & uint

+++ services/storage_manager/storage/src/storage_status_service.cpp
@@ -116,7 +116,7 @@ int32_t StorageStatusService::GetUserStorageStats(int32_t userId, StorageStats &
     storageStats.video_ = mediaVol.GetVideosSize();
     storageStats.image_ = mediaVol.GetImagesSize();
 #endif
-    int64_t prjId = userId * USER_CONST + UID_FILE_MANAGER;
+    int32_t prjId = userId * USER_CONST + UID_FILE_MANAGER;
     std::shared_ptr<StorageDaemonCommunication> sdCommunication;
     sdCommunication = DelayedSingleton<StorageDaemonCommunication>::GetInstance();
     err = sdCommunication->GetOccupiedSpace(StorageDaemon::PRJID, prjId, storageStats.file_);
